Handle wagmi inside ethereum package (defaults only)
