+++ chains/ethereum/types/apiTypes.ts
@@ -1,6 +1,6 @@
 import type {
   Chain,
-  Client,
+  ChainProviderFn,
   FetchBalanceArgs,
   FetchTokenArgs,
   FetchTransactionArgs,
@@ -14,9 +14,6 @@ import type {
   WatchReadContractResult,
   WriteContractArgs
 } from '@wagmi/core'
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type EthereumClient = Client<any, any>
 
 export interface GetDefaultConnectorsOpts {
   appName: string
@@ -116,3 +113,11 @@ export interface WaitForTransactionOpts {
   timeout?: number
   wait?: WaitForTransactionArgs['wait']
 }
+
+export interface WagmiOptions {
+  autoConnect?: boolean
+  chains?: Chain[]
+  providers?: ChainProviderFn[]
+}
+
+export type EthereumOptions = WagmiOptions | boolean
