+++ examples/react/package.json
@@ -7,7 +7,6 @@
     "build": "next build && next export"
   },
   "dependencies": {
-    "@wagmi/core": "0.5.6",
     "@web3modal/ethereum": "2.0.0-alpha.2",
     "@web3modal/react": "2.0.0-alpha.2",
     "next": "12.3.1",
