+++ packages/core/types/configTypes.ts
@@ -1,3 +1,5 @@
+import type { EthereumOptions } from '@web3modal/ethereum'
+
 export interface ConfigOptions {
   projectId: string
   theme?: 'dark' | 'light'
@@ -10,4 +12,5 @@ export interface ConfigOptions {
     | 'orange'
     | 'purple'
     | 'teal'
+  ethereum?: EthereumOptions
 }
